apply user auth roles to sites paths

+++ resources/site.py
@@ -1,6 +1,6 @@
 """Handle site related routes."""
 from email_validator import EmailNotValidError, validate_email
-from fastapi import APIRouter, Depends, HTTPException, status
+from fastapi import APIRouter, Depends, HTTPException, Request, status
 
 from config.settings import get_settings
 from database.db import database
@@ -17,7 +17,7 @@ router = APIRouter(
 @router.post(
     "/", response_model=SiteResponse, status_code=status.HTTP_201_CREATED
 )
-async def create_site(site_data: SiteRequest):
+async def create_site(site_data: SiteRequest, request: Request):
     """Create a new site."""
     slug = create_slug()
     while await get_site_by_slug(slug):
@@ -30,37 +30,71 @@ async def create_site(site_data: SiteRequest):
             status_code=status.HTTP_400_BAD_REQUEST, detail=str(e)
         )
 
-    await database.execute(Site.insert().values(**site_data.dict(), slug=slug))
+    await database.execute(
+        Site.insert().values(
+            **site_data.dict(), slug=slug, user_id=request.state.user.id
+        )
+    )
     return SiteResponse(
         name=site_data.name,
         slug=slug,
         action=f"{get_settings().base_url}/form/{slug}",
+        redirect_url=site_data.redirect_url,
     )
 
 
 @router.get("/{slug}")
-async def get_site(slug: str):
+async def get_site(slug: str, request: Request):
     """Get a site by its slug."""
     site = await get_site_by_slug(slug)
     if not site:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, detail="Site not found"
         )
-    return site
+    if (
+        site["user_id"] != request.state.user.id
+        and not request.state.user.is_admin
+    ):
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You are not authorized to view this site",
+        )
+
+    return SiteResponse(
+        name=site["name"],
+        slug=site["slug"],
+        redirect_url=site["redirect_url"],
+        action=f"{get_settings().base_url}/form/{slug}",
+    )
 
 
 @router.get("/", response_model=list[SiteList])
-async def get_sites():
+async def get_sites(request: Request):
     """Get all sites."""
-    return await database.fetch_all(Site.select())
+    if request.state.user.is_admin:
+        return await database.fetch_all(Site.select())
+    else:
+        return await database.fetch_all(
+            Site.select().where(Site.c.user_id == request.state.user.id)
+        )
 
 
 @router.delete("/{slug}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_site(slug: str):  # type: ignore
+async def delete_site(slug: str, request: Request):  # type: ignore
     """Delete a site by its slug."""
     site = await get_site_by_slug(slug)
     if not site:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, detail="Site not found"
         )
+
+    if (
+        site["user_id"] != request.state.user.id
+        and not request.state.user.is_admin
+    ):
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You are not authorized to delete this site",
+        )
+
     await database.execute(Site.delete().where(Site.c.slug == slug))
